chore(k8s): trim the RestConfig guard tests
Drop the Goexit-based recorder and its goroutine helper: SaveKubectlOptions
returns after its Fatalf, so a plain FailNow suffices, and the NoFileExists
assertion already covers the case where that return is removed.

Fold the redundant tests into the existing save_test_data_test.go. The
kubeconfig round trip duplicated TestSaveAndLoadKubectlOptions, and the
message assertions belong with the rejection test that already captures it.

Shorten the godoc and the failure message; the design rationale lives in the
PR description.
